Drive both record table drag-and-drop providers from one implementation
The grouped and ungrouped table bodies each had a 170-line dnd-kit
provider that differed in three places: where the droppable item count
comes from (the group's record ids vs the virtualized total), which
drop hook applies the move, and whether the drag overlay needs a
record-group body context. RecordTableBodyDragDropContextProvider now
takes those as getDroppableItemCount, onRecordDrop and
renderDragOverlay; the two named providers keep their call sites and
supply them.

The ungrouped provider skipped a drop landing on the source index
without comparing droppables; every ungrouped row and the end drop zone
share RECORD_TABLE_NO_RECORD_GROUP_DROPPABLE_ID, so the grouped check
(same droppable and same index) is the same condition there.
